requirements.txt
openai==2.3.0
pydantic==2.12.2
python-dotenv==1.1.1
openai-agents==0.3.3

matplotlib==3.10.7
numpy==2.3.4